@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,7 +26,8 @@ dependencies:
   cupertino_icons: ^0.1.3
 
   # To use custom icons
-  font_awesome_flutter: ^8.7.0
+  # font_awesome_flutter: ^8.7.0
+  font_awesome_flutter: ^9.1.0
 
   # Credit card validator package
   credit_card_validate:
